# Fix performance regression in amd64 decoder