In the fix for #163, I accidentally rerouted matches without literals through the slow match copy path. This undoes the mistake, with the additional optimization of using 32-bit instructions where possible, to counter some slowdown in BenchmarkUncompressDigits.

Compared to 257c664d416bd8deace75610eedccf6504312315:

    name                old speed      new speed      delta
    UncompressPg1661-8  1.19GB/s ± 1%  1.19GB/s ± 1%    ~     (p=0.198 n=10+20)
    UncompressDigits-8  2.33GB/s ± 1%  2.38GB/s ± 1%  +2.14%  (p=0.000 n=10+19)
    UncompressTwain-8   1.22GB/s ± 1%  1.23GB/s ± 1%  +0.46%  (p=0.015 n=10+20)
    UncompressRand-8    3.89GB/s ± 2%  3.88GB/s ± 2%    ~     (p=0.923 n=10+12)

Compared to 947bcf2cb7b335de36ac19425f9dbe257f85d38c:

    name                old speed      new speed      delta
    UncompressPg1661-8   758MB/s ± 1%  1189MB/s ± 1%  +56.76%  (p=0.000 n=10+20)
    UncompressDigits-8  2.32GB/s ± 2%  2.38GB/s ± 1%   +2.65%  (p=0.000 n=10+19)
    UncompressTwain-8    812MB/s ± 1%  1227MB/s ± 1%  +51.20%  (p=0.000 n=10+20)
    UncompressRand-8    3.90GB/s ± 2%  3.87GB/s ± 2%     ~     (p=0.062 n=9+20)

Also includes a fix to a benchmark.